requirements.txt
opencv-python~=4.11.0.86
numpy~=2.2.4
pytesseract~=0.3.13
pillow~=11.1.0
pip~=23.2.1
packaging~=24.2
tqdm~=4.36.1
requests~=2.22.0
pytz~=2025.2
setuptools~=41.0.1
urllib3~=2.4.0
pandas~=2.2.3
python-dateutil~=2.9.0.post0
certifi~=2025.1.31
six~=1.17.0
idna~=2.8
chardet~=3.0.4
xmljson~=0.2.0